Changed double to float to reduce size in bin files

+++ source_files/Argument_Min_Distance.cpp
@@ -4,8 +4,8 @@
 int Argument_Min_Distance(Graph *G, Query *Q, std::set<int> *Argument_Set, int x)
 {
     int min_index = -1;
-    double min_distance, temp_distance;
-    min_distance = std::numeric_limits<double>::max();
+    float min_distance, temp_distance;
+    min_distance = std::numeric_limits<float>::max();
     // Finding from the set the node with the smallest distance and returning it's location on the graph
     for (std::set<int>::iterator it = Argument_Set->begin(); it != Argument_Set->end(); it++)
     {
